refactor: rename fiop to navigation
